Add cold start detection
If a google cloud function is experiencing a cold start, TLS handshakes were observed
to sometimes take a really long time just for the first request. This function detects
cold starts by checking if the global `http.Client` `client` already exists. If not,
it will issue a GET request to the target so the subsequent http probe will not be
affected by the cold start.
https://cloud.google.com/functions/docs/concepts/exec#function_scope_versus_global_scope